Add patterns for creation-menu components
