+++ src/Button.js
@@ -4,9 +4,6 @@ import React from "react";
 import './Button.css'
 
 export default class Button extends Component{
-    constructor(props){
-        super(props);
-    }
 
     render(){
         return(
