+++ src/Input.js
@@ -2,9 +2,6 @@ import {Component} from "react";
 import React from "react";
 
 export default class Input extends Component{
-    constructor(props){
-        super(props);
-    }
 
     render() {
         return(
