+++ src/lobby/LobbyPage.js
@@ -1,19 +1,15 @@
 import React, { Component } from 'react'
 
-import RoomList from './RoomList'
+import RoomList from './rooms-list/RoomList'
+import CreationMenu from './creation-menu/CreationMenu'
 import ChatProfileBlock from './ChatProfileBlock'
 import './LobbyPage.css'
 
 export default class LobbyPage extends Component{
-    constructor(props) {
-        super(props);
-
-    }
-
     render() {
         return (
             <div id={'main-menu'}>
-                <RoomList/>
+                <CreationMenu/>
                 <ChatProfileBlock/>
             </div>
         );
