+++ src/lobby/Tab.js
@@ -3,10 +3,6 @@ import React, { Component } from 'react'
 import './Tab.css'
 
 export default class Tab extends Component {
-    constructor(props) {
-        super(props);
-
-    }
 
     render() {
         return (
