+++ src/lobby/creation-menu/CreationMenu.js
@@ -1,0 +1,16 @@
+import React, {Component} from 'react'
+
+import RoomInfo from './RoomInfo'
+import PlayersList from './PlayersList'
+import PackagesList from './PackagesList'
+import './CreationMenu.css'
+
+export default class CreationMenu extends Component{
+    render() {
+        return (
+            <div>
+
+            </div>
+        );
+    }
+}
