+++ src/lobby/creation-menu/RoomInfo.js
@@ -1,0 +1,13 @@
+import React, {Component} from 'react'
+
+import './RoomInfo.css'
+
+export default class RoomInfo extends Component{
+    render() {
+        return (
+            <div>
+
+            </div>
+        );
+    }
+}
