+++ src/lobby/user-profile/Avatar.js
